requirements.txt
Django==4.2.7
django-allauth==0.61.1
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-filter==24.1
django-widget-tweaks==1.5.0
django-mptt==0.15.0
Pillow==10.2.0
python-dotenv==1.0.1 